requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.12.0
google-cloud-bigquery==3.12.0